split the writing section to its own page and removed form homepage
